Server/QueryTool_Project/requirements.txt
django
djangorestframework
drf_spectacular
sqlalchemy
